+++ packages/acr/src/extensions/util.ts
@@ -1,6 +1,6 @@
 export function getMarketPlaceUrl(publisher: string, name: string, version: string) {
   return [
-    'kt-ext://alipay-rmsdeploy-image.cn-hangzhou.alipay.aliyun-inc.com/marketplace/assets',
-    `/${publisher}.${name}/v${version}/extension`,
+    'kt-ext://alipay-rmsdeploy-image.cn-hangzhou.alipay.aliyun-inc.com/marketplace/extension',
+    `/${publisher}.${name}-${version}`,
   ].join('');
 }