+++ packages/code-api/src/antcode/antcode.service.ts
@@ -262,6 +262,19 @@ export class AntCodeAPIService implements ICodeAPIService {
   }
 
   async searchFile(repo: IRepositoryModel, searchString: string, options: { limit: number }) {
+    // TODO：目前 OpenSumi 默认为 ''，后续得实现 queryBuilder 逻辑
+    if (searchString === '') {
+      const res = await this.request<string[]>(
+        `/api/v3/projects/${this.getProjectId(repo)}/repository/files_list`,
+        {
+          params: {
+            ref_name: repo.commit,
+          },
+        }
+      );
+      // 默认做个限制，防止请求过多
+      return res.slice(0, 1000);
+    }
     const reqRes = await this.request<string[]>(
       `/api/v3/projects/${this.getProjectId(repo)}/repository/files_search`,
       {